# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for BloodConnect donor, blood request, donation, and hospital services.",
        "title": "BloodConnect Donor Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1324",
    "basePath": "/api/v1",
    "paths": {
...
        "/blood-requests/{id}/invite/{donorId}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invite a donor to a blood request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donor Matches"
                ],
                "summary": "Invite donor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blood Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Donor ID",
                        "name": "donorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donor invited successfully",
                        "schema": {
                            "$ref": "#/definitions/helper.DonorMatchSwaggoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
...
        "/donations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a pending donation from an accepted donor match",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donations"
                ],
                "summary": "Create donation",
                "parameters": [
                    {
                        "description": "Donation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.DonationReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Donation created",
                        "schema": {
                            "$ref": "#/definitions/helper.DonationSwaggoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
        "/donations/{id}/complete": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Complete a pending donation and confirm it by the requester",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donations"
                ],
                "summary": "Complete donation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donation completed",
                        "schema": {
                            "$ref": "#/definitions/helper.DonationSwaggoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
        "/donor-matches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get donor matches from blood requests owned by the authenticated requester",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donor Matches"
                ],
                "summary": "Get donor matches for requester",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.DonorMatchListSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
        "/donor-matches/{id}/accept": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Accept donor invitation, validate blood request quantity, update distance, and create a pending donation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donor Matches"
                ],
                "summary": "Accept donor invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donor Match ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation accepted",
                        "schema": {
                            "$ref": "#/definitions/helper.DonorMatchSwaggoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
        "/donor-matches/{id}/decline": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Decline a donor match invitation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donor Matches"
                ],
                "summary": "Decline donor invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donor Match ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation declined",
                        "schema": {
                            "$ref": "#/definitions/helper.DonorMatchSwaggoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.DonationReq": {
            "type": "object",
            "properties": {
                "blood_request_id": {
                    "type": "string"
                },
                "donor_match_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.BloodRequest": {
            "type": "object",
            "properties": {
                "blood_type": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "donor_matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DonorMatch"
                    }
                },
                "hospital_external_id": {
                    "type": "string"
                },
                "hospital_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "needed_at": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "requester_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "urgency": {
                    "type": "string"
                }
            }
        },
        "entity.Donation": {
            "type": "object",
            "properties": {
                "blood_request_id": {
                    "type": "string"
                },
                "confirmed_by": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "donation_date": {
                    "type": "string"
                },
                "donor_id": {
                    "type": "string"
                },
                "donor_match_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entity.DonorMatch": {
            "type": "object",
            "properties": {
                "blood_request": {
                    "$ref": "#/definitions/entity.BloodRequest"
                },
                "blood_request_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "distance_km": {
                    "type": "number"
                },
                "donor_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "helper.DonationSwaggoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.Donation"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.DonorMatchListSwaggoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DonorMatch"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.DonorMatchSwaggoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.DonorMatch"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.ErrorDetailSwaggoResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "helper.ErrorSwaggoResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/helper.ErrorDetailSwaggoResponse"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}